client/requirements.txt
psutil
paho-mqtt==1.6.1